adjust menu func.

+++ src/main/java/com/svm/backend/admin/controller/OrganizationController.java
@@ -19,9 +19,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
-import org.springframework.data.web.PageableDefault;
-import org.springframework.data.web.SortDefault;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Controller;
 import org.springframework.util.StopWatch;
@@ -51,14 +48,23 @@ public class OrganizationController {
     @Autowired
     ApiEventRepository apiEventRepository;
 
+    /**
+     *
+     * @param request
+     * @param pageNum
+     * @param pageSize
+     * @param userAgent
+     * @param principal
+     * @return
+     */
     @RequestMapping(value = "/listD", method = RequestMethod.GET)
     @ResponseBody
     public CommonResult<CommonPage<Organization>> listOrganizationPage(
             HttpServletRequest request,
-            @RequestParam(value = "pageNum") Integer pageNum,
-            @RequestParam(value = "pageSize") Integer pageSize,
+            @RequestParam(value = "pageNum", required = false) Integer pageNum,
+            @RequestParam(value = "pageSize", required = false) Integer pageSize,
             @RequestHeader(value = "User-Agent") String userAgent,
-            Principal principal){
+            Principal principal) {
 
         StopWatch sw = new StopWatch();
         sw.start("Org listD Start");
@@ -70,7 +76,7 @@ public class OrganizationController {
                 .orElseThrow(() -> new UsernameNotFoundException("User Not Found: " + username));
 
         //正文開始
-        pageNum = pageNum -1 ;
+        pageNum = pageNum - 1;
         List<Organization> orgList = organizationRepository.findAll();
         Page<Organization> pageData = PageUtil.listToPage(orgList, pageNum, pageSize);
         sw.stop();
@@ -82,6 +88,13 @@ public class OrganizationController {
         return CommonResult.success(CommonPage.restPage(pageData));
     }
 
+    /**
+     *
+     * @param request
+     * @param principal
+     * @param userAgent
+     * @return
+     */
     @RequestMapping(value = "/treeList", method = RequestMethod.GET)
     @ResponseBody
     public CommonResult<List<OrganizationNode>> treeList(
@@ -91,21 +104,37 @@ public class OrganizationController {
 
         List<Organization> organizationList = organizationRepository.findAll();
         List<OrganizationNode> result = organizationList.stream()
-                .filter(organization -> organization.getParent_id().equals(0L))
+                .filter(organization -> organization.getParentId().equals(0L))
                 .map(organization -> covertOrgNode(organization, organizationList)).collect(Collectors.toList());
 
         return CommonResult.success(result);
     }
+
+    /**
+     *
+     * @param organization
+     * @param organizationList
+     * @return
+     */
     private OrganizationNode covertOrgNode(Organization organization, List<Organization> organizationList) {
         OrganizationNode node = new OrganizationNode();
         BeanUtils.copyProperties(organization, node);
         List<OrganizationNode> children = organizationList.stream()
-                .filter(subOrg -> subOrg.getParent_id().equals(organization.getId()))
+                .filter(subOrg -> subOrg.getParentId().equals(organization.getId()))
                 .map(subOrg -> covertOrgNode(subOrg, organizationList)).collect(Collectors.toList());
         node.setChildren(children);
         return node;
     }
 
+    /**
+     *
+     * @param request
+     * @param principal
+     * @param userAgent
+     * @param pageSize
+     * @param pageNum
+     * @return
+     */
     @RequestMapping(value = "/listLevel", method = RequestMethod.GET)
     @ResponseBody
     public CommonResult<CommonPage<Organization>> listLevel(
@@ -115,30 +144,47 @@ public class OrganizationController {
             @RequestParam(value = "pageSize", defaultValue = "5") Integer pageSize,
             @RequestParam(value = "pageNum", defaultValue = "1") Integer pageNum) {
 
-        pageNum = pageNum -1;
+        pageNum = pageNum - 1;
         Pageable pageable = PageRequest.of(pageNum, pageSize);
         Page<Organization> orgList = organizationRepository.findAll(pageable);
 
         return CommonResult.success(CommonPage.restPage(orgList));
     }
 
+    /**
+     *
+     * @param request
+     * @param principal
+     * @param userAgent
+     * @param id
+     * @return
+     */
     @RequestMapping(value = "/{id}", method = RequestMethod.GET)
     @ResponseBody
-    public CommonResult<Organization> getItem(
+    public CommonResult<Organization> getIdInfo(
             HttpServletRequest request,
             Principal principal,
             @RequestHeader(value = "User-Agent") String userAgent,
             @PathVariable Long id) {
 
         Organization organization = organizationRepository.findById(id).orElseThrow();
-        log.info("Long id:"+organization.getOrg_name());
+        log.debug("Long id:" + organization.toString());
         return CommonResult.success(organization);
 
     }
 
+    /**
+     *
+     * @param request
+     * @param principal
+     * @param userAgent
+     * @param id
+     * @param organization
+     * @return
+     */
     @RequestMapping(value = "/update/{id}", method = RequestMethod.POST)
     @ResponseBody
-    public CommonResult updateOrganization(
+    public CommonResult update(
             HttpServletRequest request,
             Principal principal,
             @RequestHeader(value = "User-Agent") String userAgent,
@@ -149,9 +195,18 @@ public class OrganizationController {
         return CommonResult.success(null);
     }
 
+    /**
+     *
+     * @param request
+     * @param userAgent
+     * @param principal
+     * @param id
+     * @param status
+     * @return
+     */
     @RequestMapping(value = "/updateStatus/{id}", method = RequestMethod.POST)
     @ResponseBody
-    public CommonResult updateOrganizationStatus(
+    public CommonResult updateStatus(
             HttpServletRequest request,
             @RequestHeader(value = "User-Agent") String userAgent,
             Principal principal,
@@ -166,35 +221,54 @@ public class OrganizationController {
 
     }
 
+    /**
+     *
+     * @param request
+     * @param principal
+     * @param userAgent
+     * @param organization
+     * @return
+     */
     @RequestMapping(value = "/create", method = RequestMethod.POST)
     @ResponseBody
-    public CommonResult createOrganization(
+    public CommonResult create(
             HttpServletRequest request,
             Principal principal,
             @RequestHeader(value = "User-Agent") String userAgent,
             @RequestBody Organization organization) {
 
-        Organization upper_org = organizationRepository.findById(organization.getParent_id()).orElseThrow();
+        Organization upper_org = organizationRepository.findById(organization.getParentId()).orElseThrow();
         organization.setCreateTime(new Date());
-        organization.setLevel(upper_org.getLevel()+1);
+        organization.setLevel(upper_org.getLevel() + 1);
         organizationRepository.save(organization);
         return CommonResult.success(null);
     }
 
+    /**
+     *
+     * @param request
+     * @param principal
+     * @param userAgent
+     * @param id
+     * @return
+     */
     @RequestMapping(value = "/delete/{id}", method = RequestMethod.POST)
     @ResponseBody
-    public CommonResult deleteOrganization(
+    public CommonResult delete(
             HttpServletRequest request,
             Principal principal,
             @RequestHeader(value = "User-Agent") String userAgent,
             @PathVariable Long id) {
 
         Organization organization = organizationRepository.findById(id).orElseThrow();
+        List<User> userList = userRepository.findUserByOrgId(id);
         List<Organization> organizationList = organizationRepository.getOrgListByParentId(id);
-        if(organizationList.isEmpty()){
+        if (organizationList.isEmpty() && userList.isEmpty()) {
             organizationRepository.deleteById(id);
-        }else{
-            return CommonResult.failed("有下層組織，無法刪除");
+        } else {
+            log.debug("delete org "+organization.getOrgName()+" ,but userlist size:"+userList.size());
+            log.debug("delete org "+organization.getOrgName()+",but orglist size:"+organizationList.size());
+            return CommonResult.failed(organization.getOrgName() + "有下層組織或帳號，無法刪除");
         }
         return CommonResult.success(null);
     }
